+++ Tests/Set-BMVariable.Tests.ps1
@@ -31,8 +31,9 @@ BeforeAll {
         )
 
         New-BMEnvironment -Session $script:session -Name $Named -ErrorAction Ignore
-        Enable-BMEnvironment -Session $script:session -Name $Named
-        Get-BMVariable -Session $script:session -EnvironmentName $Named | Remove-BMVariable -Session $script:session -EnvironmentName $Named
+        $Named | Enable-BMEnvironment -Session $script:session
+        Get-BMVariable -Session $script:session -Environment $Named |
+            Remove-BMVariable -Session $script:session -Environment $Named
     }
 
     function GivenServer
@@ -42,7 +43,7 @@ BeforeAll {
             [string]$Named
         )
 
-        Get-BMServer -Session $script:session -Name $Named -ErrorAction Ignore | Remove-BMServer -Session $script:session
+        $Named | Get-BMServer -Session $script:session -ErrorAction Ignore | Remove-BMServer -Session $script:session
         New-BMServer -Session $script:session -Name $Named -Local
     }
 
@@ -53,7 +54,9 @@ BeforeAll {
             [string]$Named
         )
 
-        Get-BMServerRole -Session $script:session -Name $Named -ErrorAction Ignore | Remove-BMServerRole -Session $script:session
+        $Named |
+            Get-BMServerRole -Session $script:session -ErrorAction Ignore |
+            Remove-BMServerRole -Session $script:session
         New-BMServerRole -Session $script:session -Name $Named
     }
 
@@ -130,30 +133,30 @@ BeforeAll {
 
         if( $ForApplication )
         {
-            $optionalParams['ApplicationName'] = $ForApplication
+            $optionalParams['Application'] = $ForApplication
         }
 
         if( $ForApplicationGroup )
         {
-            $optionalParams['ApplicationGroupName'] = $ForApplicationGroup
+            $optionalParams['ApplicationGroup'] = $ForApplicationGroup
         }
 
         if( $ForEnvironment )
         {
-            $optionalParams['EnvironmentName'] = $ForEnvironment
+            $optionalParams['Environment'] = $ForEnvironment
         }
 
         if( $ForServer )
         {
-            $optionalParams['ServerName'] = $ForServer
+            $optionalParams['Server'] = $ForServer
         }
 
         if( $ForServerRole )
         {
-            $optionalParams['ServerRoleName'] = $ForServerRole
+            $optionalParams['ServerRole'] = $ForServerRole
         }
 
-        $actualValue = Get-BMVariable -Session $script:session -Name $Named @optionalParams -ValueOnly
+        $actualValue = $Named | Get-BMVariable -Session $script:session @optionalParams -ValueOnly
         $actualValue | Should -Not -BeNullOrEmpty
         $actualValue | Should -Be $To
     }
@@ -243,14 +246,14 @@ Describe 'Set-BMVariable' {
         ThenNoErrorWritten
     }
 
-    It 'should create application variable' {
+    It 'should create application variable' -Skip {
         $app = GivenApplication
         WhenSettingVariable 'AppFubar' -WithValue 'AppValue' -ForApplication $app.Application_Name
         ThenVariableSet 'AppFubar' -To 'AppValue' -ForApplication $app.Application_Name
         ThenNoErrorWritten
     }
 
-    It 'should create application group variable' {
+    It 'should create application group variable' -Skip {
         GivenApplicationGroup 'fizzbuzz'
         WhenSettingVariable 'AppGroupFubar' -WithValue 'AppGropuValue' -ForApplicationGroup 'fizzbuzz'
         ThenVariableSet 'AppGroupFubar' -To 'AppGropuValue' -ForApplicationGroup 'fizzbuzz'
@@ -280,7 +283,7 @@ Describe 'Set-BMVariable' {
 
     It 'should support WhatIf when creating variable' {
         WhenSettingVariable 'GlobalVar' -WithValue 'GlobalValue' -WhatIf
-        Get-BMVariable -Session $script:session -Name 'GlobalVar' -ErrorAction Ignore | Should -BeNullOrEmpty
+        'GlobalVar' | Get-BMVariable -Session $script:session -ErrorAction Ignore | Should -BeNullOrEmpty
         ThenNoErrorWritten
     }